Rebalance between lib and exe
