+++ src/GHCMod.hs
@@ -20,6 +20,7 @@ import System.IO
 import System.Exit
 import Text.PrettyPrint hiding ((<>))
 import Language.Haskell.GhcMod.Options.Options
+import GHCMod.Options
 import Prelude
 
 ghcModStyle :: Style
