+++ src/GHCMod/Options/ShellParse.hs
@@ -13,7 +13,7 @@
 --
 -- You should have received a copy of the GNU Affero General Public License
 -- along with this program.  If not, see <http://www.gnu.org/licenses/>.
-module Language.Haskell.GhcMod.Options.ShellParse (parseCmdLine) where
+module GHCMod.Options.ShellParse (parseCmdLine) where
 
 import Data.Char
 import Data.List
