+++ src/GHCMod/Version.hs
@@ -14,7 +14,7 @@
 -- You should have received a copy of the GNU Affero General Public License
 -- along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-module Language.Haskell.GhcMod.Options.Version where
+module GHCMod.Version where
 
 import Paths_ghc_mod
 import Data.Version (showVersion)
